Changed demo package names.
