+++ examples/centralized/src/main/java/co/com/bancolombia/binstash/democentralized/CacheDemoCentralizedApplication.java
@@ -1,0 +1,11 @@
+package co.com.bancolombia.binstash.democentralized;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CacheDemoCentralizedApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(CacheDemoCentralizedApplication.class, args);
+    }
+}
