+++ examples/centralized/src/main/java/co/com/bancolombia/binstash/democentralized/model/Address.java
@@ -1,0 +1,13 @@
+package co.com.bancolombia.binstash.democentralized.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Address {
+    private String street;
+    private String city;
+}
