+++ examples/centralized/src/main/java/co/com/bancolombia/binstash/democentralized/model/DummyRepo.java
@@ -1,0 +1,7 @@
+package co.com.bancolombia.binstash.democentralized.model;
+
+import reactor.core.publisher.Mono;
+
+public interface DummyRepo {
+    Mono<Person> findByName(String name);
+}
