+++ examples/centralized/src/main/java/co/com/bancolombia/binstash/democentralized/model/Person.java
@@ -1,0 +1,13 @@
+package co.com.bancolombia.binstash.democentralized.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Person {
+    private String name;
+    private Address address;
+}
