--- examples/centralized/src/main/java/co/com/bancolombia/binstash/example/model/DummyRepo.java
@@ -1,7 +1,0 @@
-package co.com.bancolombia.binstash.example.model;
-
-import reactor.core.publisher.Mono;
-
-public interface DummyRepo {
-    Mono<Person> findByName(String name);
-}
