+++ examples/hybrid/src/main/java/co/com/bancolombia/binstash/demohybrid/CacheDemoHybridApplication.java
@@ -1,0 +1,11 @@
+package co.com.bancolombia.binstash.demohybrid;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CacheDemoHybridApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(CacheDemoHybridApplication.class, args);
+    }
+}
