+++ examples/hybrid/src/main/java/co/com/bancolombia/binstash/demohybrid/model/Address.java
@@ -1,0 +1,13 @@
+package co.com.bancolombia.binstash.demohybrid.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Address {
+    private String street;
+    private String city;
+}
