+++ examples/hybrid/src/main/java/co/com/bancolombia/binstash/demohybrid/model/Person.java
@@ -1,0 +1,13 @@
+package co.com.bancolombia.binstash.demohybrid.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Person {
+    private String name;
+    private Address address;
+}
