--- examples/hybrid/src/main/java/co/com/bancolombia/binstash/example/CacheDemoHybridApplication.java
@@ -1,11 +1,0 @@
-package co.com.bancolombia.binstash.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CacheDemoHybridApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(CacheDemoHybridApplication.class, args);
-    }
-}
