--- examples/hybrid/src/main/java/co/com/bancolombia/binstash/example/model/Person.java
@@ -1,13 +1,0 @@
-package co.com.bancolombia.binstash.example.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Person {
-    private String name;
-    private Address address;
-}
