+++ examples/local/src/main/java/co/com/bancolombia/binstash/demolocal/CacheDemoLocalApplication.java
@@ -1,0 +1,11 @@
+package co.com.bancolombia.binstash.demolocal;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CacheDemoLocalApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(CacheDemoLocalApplication.class, args);
+    }
+}
