--- examples/local/src/main/java/co/com/bancolombia/binstash/example/CacheDemoLocalApplication.java
@@ -1,11 +1,0 @@
-package co.com.bancolombia.binstash.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CacheDemoLocalApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(CacheDemoLocalApplication.class, args);
-    }
-}
