--- examples/local/src/main/java/co/com/bancolombia/binstash/example/model/Address.java
@@ -1,13 +1,0 @@
-package co.com.bancolombia.binstash.example.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Address {
-    private String street;
-    private String city;
-}
